break page components into different components
