+++ apps/nameguard.io/src/app/components/MobileSectionDivider.tsx
@@ -1,0 +1,7 @@
+export function MobileSectionDivider() {
+  return (
+    <div className="flex gt_mobile:hidden items-center justify-center w-full h-fit px-5">
+      <span className="bg-gray-200 h-[1px] w-full"></span>
+    </div>
+  );
+}
