+++ apps/nameguard.io/src/types/imageTypes.ts
@@ -1,0 +1,6 @@
+export type ImageCharacteristics = {
+  source: string;
+  tagWidth: number;
+  tagHeight: number;
+  styles?: string;
+};
